week_7/CrewAI Agents - Google Cloud Run Deployment/requirements.txt
streamlit
crewai
langchain-google-genai
python-dotenv